Fix --matchdirs listing of children for filtered dirs
Using `-P pattern --matchdirs` currently only lists all children if a
match is at the first level. This is not expected, Unix tree will show
all children if any directory on the tree matches.

This change removes a buggy depth check to allow listing of all children
on matching directory.